PLAT-763 Auto-Complete inputs: Do not infer entries from ID substrings
Fixes a langstanding defect which caused that an entry could be inferred from a partial ID substring which, in turn, could not be resolved to an existing entry on its own.

For example, an existing entry with ID 23 was inferred from input text `"2"` (or `"3"`) if there was no entry with ID 2 (or 3).

Semi-related cleanup: Merged `AbstractEmfEntityInputEngine` into `EmfEntityInputEngine` (that base class was pointless).