PLAT-1136-2 Implement setPreventDefaultOnWheel
We currently cannot write automated tests for this. We could try https://www.selenium.dev/documentation/webdriver/actions_api/wheel/ but that would require a switch to Chrome based Selenium nodes. So, this PR comes with a manual test plan only.

There are 12 events in this GIF:
- 3x WHEEL_DOWN (default behavior allowed = scrolling)
- 3x Shift + WHEEL_DOWN (default behavior suppressed = NO scrolling)
- 3x WHEEL_DOWN (default behavior allowed = scrolling)
- 3x Shift + WHEEL_DOWN (default behavior suppressed = NO scrolling)

This is the wheel test element that can be seen in the GIF:
```
private class WheelTestDiv extends DomDiv implements IDomWheelEventHandler {

	public WheelTestDiv() {

		appendText("wheel me");
		setStyle(CssStyle.BACKGROUND_COLOR, "yellow");

		CurrentDomDocument.get().getEngine().setPreventDefaultOnWheel(this, Set.of(DomModifier.CONTROL), true);
	}

	@Override
	public void handleWheel(IDomEvent event) {

		if (event.getDeltaY() > 0) {
			Log.finfo("down");
			appendChild(new DomDiv()).appendText("down");
			Sleep.sleep(500);
		} else if (event.getDeltaY() < 0) {
			Log.finfo("up");
			appendChild(new DomDiv()).appendText("up");
			Sleep.sleep(500);
		}
	}
}
```