EXP-2296-5 Table export: Simplify exception handling
Before this PR, we tried really hard to show well-translated but totally technical and overly specific exception messages to users, about which they could do absolutely nothing.
Also, there were scenarios in which we only showed the fallback exception message.
This was simplified and unified, such that now merely show "Table export failed." to the user when anything goes wrong.